requirements.txt
playwright==1.40.0
beautifulsoup4~=4.13.4
google-generativeai==0.3.2
python-dotenv==1.0.0
neo4j==5.15.0
pydantic~=2.11.7
fastapi~=0.116.1
langchain-core~=0.3.72
langchain-openai~=0.3.28
openai~=1.97.1
requests~=2.32.4
websockets~=15.0.1